added react-router-dom and implemented basic routing
